server/agent/graph_config.py
from dotenv import load_dotenv 
from neo4j import AsyncGraphDatabase
from py2neo import Graph
import os
load_dotenv()



# ============= CONFIGURATION =============
NEO4J_URI = os.getenv("NEO4J_URI")
NEO4J_USER = os.getenv("NEO4J_USER")
NEO4J_PASSWORD = os.getenv("NEO4J_PASSWORD")
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")

driver = AsyncGraphDatabase.driver(NEO4J_URI, auth=(NEO4J_USER, NEO4J_PASSWORD))

