server/agent/requirements.txt

langchain==1.0.3
chromadb==1.3.0
unstructured[all_docs]==0.18.15
langchain-chroma==1.0.0
langchain-core==1.0.3
langchain-huggingface==1.0.1
transformers==4.57.1
sentence-transformers==5.1.2
torch==2.9.0
python-dotenv==1.2.1
langchain-google-genai==3.0.1
langchain-openai==1.0.2

langchain-text-splitters==1.0.0
langchain-classic==1.0.0
langchain-deepseek==1.0.0
langchain-community==0.4.1 
langchain-ollama==1.0.0
langchain_groq==1.0.0
spacy==3.8.11 
redis==7.0.1
pymupdf==1.26.6
py2neo==2021.2.4
neo4j==6.0.3
asyncpg==0.31.0
uvicorn 
fastapi
sqlalchemy==4.15.0
passlib==1.7.4
alembic==1.17.2
